+++ src/spyglass/utils/dj_helper_fn.py
@@ -560,6 +560,8 @@ def make_file_obj_id_unique(nwb_path: str):
         with h5py.File(nwb_path, "a") as f:
             f.attrs["object_id"] = new_id
     except (BlockingIOError, OSError):
+        from spyglass.common.common_usage import ExportErrorLog
+
         ExportErrorLog().insert1(
             {
                 "file": nwb_path,
@@ -738,36 +740,3 @@ def _replace_nan_with_default(data_dict, default_value=-1.0):
             result[key] = default_value
 
     return result
-
-
-# Log table for errors encountered during file edits
-schema = dj.schema("common_export_error_log")
-
-
-@schema
-class ExportErrorLog(dj.Manual):
-    definition = """
-    file: varchar(255)  # file being processed
-    source: varchar(255)  # source of the error (e.g., table name or function)
-    ---
-    """
-
-    @staticmethod
-    def _logger_warning(key):
-        logger.warning(
-            f"Logging export error for file: {key.get('file', 'unknown')}"
-            + f" from source: {key.get('source', 'unknown')}"
-        )
-
-    def insert1(self, key, **kwargs):
-        """Insert a new entry into the ExportErrorLog table.
-
-        Parameters
-        ----------
-        key : dict
-            Dictionary containing the primary key fields for the table.
-        **kwargs : dict
-            Additional keyword arguments for non-primary key fields.
-        """
-        self._logger_warning(key)
-        super().insert1(key, **kwargs)